test(lossless): the GraphQLx wiring guards, in a file a GraphQLx build compiles
`tests/lossless_wiring.rs` is headed `#![cfg(all(feature = "rowan", feature = "graphql"))]`
and held four GraphQLx items behind an inner `#[cfg(feature = "graphqlx")]`: the `const _`
`SURFACES_TRIVIA` assertion for the GraphQLx lossless token, its named companion,
`the_graphqlx_lossless_lexer_surfaces_every_trivia_form`, and
`the_graphqlx_lossless_suite_exposes_its_three_roots`. An inner `cfg` under a header that
already demands the other dialect reads as "compiled when both are on", and it is — but the
build these guards exist for is `--features "rowan,graphqlx"`, where the header is false and
all four are compiled away.

Nothing said so, because until the previous commit no such build existed: the self
dev-dependency pulled `graphql` in from the library's defaults, so every test target had it
and the header could not be wrong. It can now, and this is what it costs:

    --features "std,parser,rowan,graphqlx" --test lossless_wiring     0 tests    (was 7 before #136)
    --features "std,parser,rowan,graphqlx" --test lossless_x_wiring   3 tests
    --features "std,parser,rowan,graphql"  --test lossless_wiring     4 tests
    --features "std,parser,rowan,graphql"  --test lossless_x_wiring   0 tests

`4 + 3` is the seven the one file used to report.

`tests/lossless_x_wiring.rs` takes the header the other fifteen `lossless_x_*` files use,
verbatim. The inner `#[cfg]`s are gone from both halves rather than moved with the items: a
`cfg` that restates its own file header is exactly what hides a header from the compiler,
and a header that cannot be wrong is the defect this pair is about.

`TRIVIA_PROBE`, `scan` and `assert_trivia_survives_lexing` are shared, so they move to
`tests/support/trivia_probe.rs` and both files reach them by `#[path] mod`. That is the
mechanism already in the tree for exactly this case: `tests/support/span_extent.rs` is
shared by `syntactic_span_extent.rs` and `syntactic_x_span_extent.rs`, and
`tests/support/graphqlx_padding.rs` by the two GraphQLx padding gates, each for the reason
that applies here — an integration test is its own crate, so no assertion in one file can
name an item in the other and two copies would be free to drift.

`tokoras_cst_layer_is_in_scope` deliberately does not move or gain a twin. It is a claim
about the `rowan` feature reaching tokora with no dialect in it, so a second copy under a
second header would be two statements of one fact rather than a second measurement.
